pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "git-of-theseus"
version = "0.3.4"
description = "Plot stats on Git repositories"
authors = [{ name = "Erik Bernhardsson", email = "mail@erikbern.com" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
# Core analysis dependencies. Plotting and progress-bar deps are split out into
# the `plot` extra so the package can be installed in dependency-light
# environments such as Pyodide / WASM browser builds.
dependencies = [
    "gitpython",
    "pygments",
    "wcmatch",
]

[project.scripts]
git-of-theseus-analyze = "git_of_theseus.analyze:analyze_cmdline"
git-of-theseus-survival-plot = "git_of_theseus:survival_plot_cmdline"
git-of-theseus-stack-plot = "git_of_theseus:stack_plot_cmdline"
git-of-theseus-line-plot = "git_of_theseus:line_plot_cmdline"

[project.optional-dependencies]
# Required to run the *_plot CLIs and to get a tqdm progress bar in the CLI.
plot = [
    "matplotlib",
    "numpy",
    "python-dateutil",
    "scipy",
    "tqdm",
]
dev = [
    "pytest",
    "matplotlib",
    "numpy",
    "python-dateutil",
    "scipy",
    "tqdm",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.urls]
Homepage = "https://github.com/erikbern/git-of-theseus"
